# Conflicts:
#	CHANGES.md
#	multiply_data_access/version.py

+++ multiply_data_access/aws_s2_file_system.py
@@ -108,7 +108,8 @@ class AwsS2FileSystem(LocallyWrappedFileSystem):
         return parameters
 
     def clear_cache(self):
-        shutil.rmtree(self._temp_dir)
+        if os.path.exists(self._temp_dir):
+            shutil.rmtree(self._temp_dir)
 
 
 class AwsS2FileSystemAccessor(FileSystemAccessor):
