+++ multiply_data_access/general_remote_access.py
@@ -182,7 +182,8 @@ class HttpFileSystem(LocallyWrappedFileSystem):
         return parameters
 
     def clear_cache(self):
-        shutil.rmtree(self._temp_dir)
+        if os.path.exists(self._temp_dir):
+            shutil.rmtree(self._temp_dir)
 
 
 class HttpFileSystemAccessor(FileSystemAccessor):
