+++ multiply_data_access/lpdaac_data_access.py
@@ -231,7 +231,8 @@ class LpDaacFileSystem(LocallyWrappedFileSystem):
         return parameters
 
     def clear_cache(self):
-        shutil.rmtree(self._temp_dir)
+        if os.path.exists(self._temp_dir):
+            shutil.rmtree(self._temp_dir)
 
 
 class LpDaacFileSystemAccessor(FileSystemAccessor):
